@@ -1,6 +1,16 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-
 import { StatusComponent } from './status.component';
+import { TouchPointService } from 'src/services/touch-point.service';
+import { of } from 'rxjs';
+
+const dummyTouchPointService: jasmine.SpyObj<TouchPointService> = jasmine.createSpyObj('TouchPointService', [
+  'getTouchPoints',
+  'getHistory',
+  'getAddress',
+  'getFrameNumber',
+  'isConnected'
+]
+);
 
 describe('StatusComponent', () => {
   let component: StatusComponent;
@@ -8,9 +18,16 @@ describe('StatusComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ StatusComponent ]
+      declarations: [ StatusComponent ],
+      providers: [
+        { provide: TouchPointService, useValue: dummyTouchPointService }
+      ]
     })
     .compileComponents();
+
+    dummyTouchPointService.getTouchPoints.and.returnValue(of([]));
+    dummyTouchPointService.isConnected.and.returnValue(true);
+    dummyTouchPointService.getFrameNumber.and.returnValue(of(1));
   });
 
   beforeEach(() => {
