@@ -20,16 +20,16 @@ describe('AppComponent', () => {
     expect(app).toBeTruthy();
   });
 
-  it(`should have as title 'reflex-emulator'`, () => {
+  it(`should have as title 'ReFlex | Emulator'`, () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.componentInstance;
-    expect(app.title).toEqual('reflex-emulator');
+    expect(app.title).toEqual('ReFlex | Emulator');
   });
 
-  it('should render title', () => {
+  it('should have content', () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.nativeElement;
-    expect(compiled.querySelector('.content span').textContent).toContain('reflex-emulator app is running!');
+    expect(compiled.querySelector('.content')).toBeTruthy();
   });
 });
